Move factory classes to tests
